+++ internal/infrastructure/datasource/entity/prepared_experiment_entity.go
@@ -8,7 +8,7 @@ import (
 
 type PreparedExperimentEntity struct {
 	ID           uint    `gorm:"primaryKey"`
-	UserID       uint    `gorm:"not null;index"`
+	UserID       uint    `gorm:"not null;default:1;index"`
 	ExperimentID uint    `gorm:"not null;index"`
 	CropAlt      float64 `gorm:"not null"`
 	BGRType      string  `gorm:"size:20;not null"`